requirements.txt
watchdog
requests[socks]
python-dotenv
pandas
openpyxl
